# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    - sku
    type: object
...
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.CategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductImageResponse'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
  github_com_abhilashdk2016_golang-ecommerce_internal_dto.UpdateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
...
  github_com_abhilashdk2016_golang-ecommerce_internal_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.PaginationMeta'
      success:
        type: boolean
    type: object
  github_com_abhilashdk2016_golang-ecommerce_internal_utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: no-email@no-email
    name: Abhilash D K
    url: http://linkedin.com/in/abhilashdk
  description: A modern e-commerce API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Retrieve all active categories
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.CategoryResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category (Admin only)
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Delete a category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Category deleted successfully
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category (Admin only)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.CategoryResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /products:
    get:
      description: Retrieve paginated list of active products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product (Admin only)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      description: Delete a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: Retrieve detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductResponse'
              type: object
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      summary: Get a product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_dto.ProductResponse'
              type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
  /products/{id}/images:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid request or file
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_abhilashdk2016_golang-ecommerce_internal_utils.Response'
      security:
      - BearerAuth: []
      summary: Upload product image
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"